ShoutWikiAds: version 0.4.3 -- add support for the Refreshed skin
The sidebar slot naming is somewhat of a clusterfuck -- the sidebar
([[MediaWiki:Sidebar]]) as well as extension-controlled areas (NewsBox)
and interlanguage links are placed in the sidebar area, div#sidebar. We
want a 125x125px ad there, just like in MonoBook et al. since it looks
good there. The slot is internally referred to as 'toolbox-button' even
though the module has to be named ext.ShoutWikiAds.refreshed.sidebar for
it to be loaded. This is confusing and messy and ugly, but at least it
works.

The other slots are leaderboards, one in the sitenotice area
('leaderboard', handled by the onSkinAfterContent hook) and another in the
footer area ('leaderboard-footer', handled by the onRefreshedFooter hook.

The aforementioned "toolbox"/sidebar ad slot requires
https://gerrit.wikimedia.org/r/301579 / the newest version of Refreshed.

ShoutWiki bug T99, ShoutWiki SVN r3930

Change-Id: I215b877123161cab81b82d07bf6bb3466d070106